refactor Add event action
